+++ js/main.js
@@ -1,3 +1,5 @@
 import {getOffers} from './data.js';
 
-getOffers(10);
+const SIMILAR_OFFER_COUNT = 10;
+
+getOffers(SIMILAR_OFFER_COUNT);
